requirements.txt
requests~=2.32.3
gradio~=5.29.1
setuptools~=80.7.1
pydantic~=2.10.6
SQLAlchemy~=2.0.41
pytest~=8.3.5
simplejson~=3.20.1
tqdm~=4.67.1
PyYAML~=6.0.2
matplotlib~=3.8.4
pandas~=2.2.2
pyperclip~=1.8.2
